acp-emergent/observer/requirements.txt
grpcio>=1.54.0
grpcio-tools>=1.54.0
protobuf>=4.22.3
pyyaml>=6.0
requests>=2.31.0
urllib3>=2.0.0
